Move logo components
The logo components were previously nested under `icons/`.
This moves them to a dedicated `logos/` directory under `components/`.
